Conflicts:
	.gitignore
	Makefile
	bauer/brueckenbauer.cc
	bauer/brueckenbauer.h
	bauer/fabrikbauer.cc
	bauer/hausbauer.cc
	bauer/vehikelbauer.cc
	bauer/wegbauer.cc
	besch/haus_besch.h
	besch/reader/building_reader.cc
	besch/vehikel_besch.h
	besch/writer/building_writer.cc
	besch/writer/vehicle_writer.cc
	boden/wege/weg.cc
	dataobj/einstellungen.cc
	dataobj/einstellungen.h
	dataobj/loadsave.cc
	dataobj/loadsave.h
	dataobj/network_cmd.h
	dataobj/network_cmd_ingame.cc
	dataobj/route.cc
	dataobj/umgebung.h
	dings/baum.cc
	dings/gebaeude.cc
	dings/leitung2.cc
	gui/banner.cc
	gui/climates.cc
	gui/convoi_info_t.cc
	gui/depot_frame.cc
	gui/fahrplan_gui.cc
	gui/fahrplan_gui.h
	gui/loadsave_frame.cc
	gui/money_frame.cc
	gui/pakselector.cc
	gui/pakselector.h
	gui/player_frame_t.cc
	gui/savegame_frame.cc
	gui/savegame_frame.h
	gui/settings_stats.cc
	gui/welt.cc
	player/ai_goods.cc
	player/simplay.cc
	player/simplay.h
	scrolltext.h
	simcity.cc
	simconvoi.cc
	simconvoi.h
	simdepot.cc
	simdepot.h
	simfab.cc
	simhalt.cc
	simlinemgmt.cc
	simmain.cc
	simsys_s.cc
	simutrans/config/simuconf.tab
	simutrans/copyright.txt
	simutrans/text/cz/new_world.txt
	simutrans/text/en/server.txt
	simversion.h
	simwerkz.cc
	simworld.cc
	simworld.h
	tpl/HOT_queue2_tpl.h
	tpl/HOT_queue_tpl.h
	tpl/prioqueue_tpl.h
	tpl/sorted_heap_tpl.h
	tpl/vector_tpl.h
	vehicle/simvehikel.cc
	vehicle/simvehikel.h

+++ gui/halt_list_frame.cc
@@ -13,12 +13,10 @@
 #include "halt_list_frame.h"
 #include "halt_list_filter_frame.h"
 
-#include "../simskin.h"
 #include "../simhalt.h"
 #include "../simware.h"
 #include "../simfab.h"
 #include "../simwin.h"
-#include "../simcolor.h"
 #include "../besch/skin_besch.h"
 
 #include "../bauer/warenbauer.h"
@@ -392,7 +390,7 @@ bool halt_list_frame_t::action_triggered( gui_action_creator_t *komp,value_t /* 
 		}
 		else {
 			filter_frame = new halt_list_filter_frame_t(m_sp, this);
-			create_win(filter_frame, w_info, (long)this);
+			create_win(filter_frame, w_info, (ptrdiff_t)this);
 		}
 	}
 	return true;
